Fix all tests so they actually pass
Primarily, this is just fixing import paths and some minor changes to function signatures and mocking behavior.

The most notable change is the move of the live tests to a new integration test category, as they are not true unit tests and require live connections to the services and valid API keys to run.  In theory, these should be fully mocked as well, but thats outside the scope of my efforts here.